+++ mvp-widget-sample/src/main/java/ru/surfstudio/android/mvpwidget/sample/ui/screen/main/widget/relative/RelativeViewConfigurator.kt
@@ -1,0 +1,29 @@
+package ru.surfstudio.android.mvpwidget.sample.ui.screen.main.widget.relative
+
+import dagger.Component
+import ru.surfstudio.android.core.mvp.configurator.ScreenComponent
+import ru.surfstudio.android.dagger.scope.PerScreen
+import ru.surfstudio.android.sample.dagger.ui.base.configurator.DefaultWidgetScreenConfigurator
+import ru.surfstudio.android.sample.dagger.ui.base.dagger.activity.DefaultActivityComponent
+import ru.surfstudio.android.sample.dagger.ui.base.dagger.widget.DefaultWidgetScreenModule
+
+/**
+ * Простой пример конфигуратора для виджета
+ */
+class RelativeViewConfigurator : DefaultWidgetScreenConfigurator() {
+
+    override fun createScreenComponent(
+            parentComponentDefault: DefaultActivityComponent,
+            widgetScreenModule: DefaultWidgetScreenModule
+    ): ScreenComponent<*> {
+        return DaggerRelativeViewConfigurator_RelativeViewComponent.builder()
+                .defaultActivityComponent(parentComponentDefault)
+                .defaultWidgetScreenModule(widgetScreenModule)
+                .build()
+    }
+
+    @PerScreen
+    @Component(dependencies = [DefaultActivityComponent::class],
+            modules = [DefaultWidgetScreenModule::class])
+    interface RelativeViewComponent : ScreenComponent<RelativeWidgetView>
+}