+++ push/src/main/java/ru/surfstudio/android/notification/ui/notification/strategies/PushHandleStrategy.kt
@@ -19,14 +19,15 @@ import android.app.Activity
 import android.app.NotificationChannel
 import android.app.PendingIntent
 import android.content.Context
+import android.widget.RemoteViews
 import androidx.core.app.NotificationCompat
 import androidx.annotation.StringRes
-import android.widget.RemoteViews
 import ru.surfstudio.android.core.ui.navigation.activity.route.ActivityRoute
 import ru.surfstudio.android.notification.R
 import ru.surfstudio.android.notification.interactor.push.BaseNotificationTypeData
 import ru.surfstudio.android.notification.interactor.push.PushInteractor
 import ru.surfstudio.android.notification.ui.notification.NotificationCreateHelper
+import ru.surfstudio.android.notification.ui.notification.groups.NotificationsGroup
 
 /**
  * Стратегия обработки пуша
@@ -37,10 +38,23 @@ import ru.surfstudio.android.notification.ui.notification.NotificationCreateHelp
  * Для Android O необходимо указать id канала
  */
 abstract class PushHandleStrategy<out T : BaseNotificationTypeData<*>> {
+
     /**
-     * тип данных пуша [BaseNotificationTypeData]
+     * Идентификатор пуш-нотификации. При совпадении идентификаторов пуш-нотификации заменяют друг
+     * друга. Если в стратегии не переопределить данное свойство, все нотификации будут
+     * автоматически получать уникальные идентификаторы.
+     */
+    open var pushId: Int = -1
+
+    /**
+     * Данные пуш-нотификации [BaseNotificationTypeData].
      */
     abstract val typeData: T
+
+    /**
+     * Группа пуш-нотификации.
+     */
+    open val group: NotificationsGroup? = null
 
     /**
      * Id канала пуш уведомлений(строковый ресурс)
@@ -79,9 +93,14 @@ abstract class PushHandleStrategy<out T : BaseNotificationTypeData<*>> {
     var channel: NotificationChannel? = null
 
     /**
-     * Кастомный builder для пушей(если нужно какое-то особое поведение)
+     * Кастомный builder для нотификаций.
      */
     var notificationBuilder: NotificationCompat.Builder? = null
+
+    /**
+     * Кастомный builder для заголовка группы нотификаций.
+     */
+    var groupSummaryNotificationBuilder: NotificationCompat.Builder? = null
 
     /**
      * Действия при нажатии на пуш
@@ -91,38 +110,51 @@ abstract class PushHandleStrategy<out T : BaseNotificationTypeData<*>> {
     /**
      * Требуемое действие нотификации
      *
-     * @param context      текущий контекст
-     * @param pushInteractor пуш интерактор
-     * @param title          заголвок пуша
-     * @param body           сообщение пуша
-     * @param data           данные пуша
+     * @param context текущий контекст
+     * @param pushInteractor пуш-интерактор
+     * @param title заголовок пуш-нотификации
+     * @param body текст пуш-нотификации
      */
     open fun handle(
             context: Context,
             pushInteractor: PushInteractor,
+            uniqueId: Int,
             title: String,
             body: String
     ) {
-
         pendingIntent = preparePendingIntent(context, title)
         notificationBuilder = makeNotificationBuilder(context, title, body)
+        groupSummaryNotificationBuilder = makeGroupSummaryNotificationBuilder(context, title, body)
         channel = makeNotificationChannel(context, title)
+        makePushId(uniqueId)
 
         pushInteractor.onNewNotification(typeData)
 
-        when (context) {
-            is Activity -> if (!handlePushInActivity(context)) {
-                showNotification(context, title, body)
-            }
+        if (context is Activity && handlePushInActivity(context)) return
+        showNotification(context, pushId, title, body)
+    }
 
-            else -> showNotification(context, title, body)
+    /**
+     * Инициализация идентификатора пуш-нотификации.
+     *
+     * Если ID пуша не задано явно в стратегии, все пуши будут иметь уникальный ID.
+     *
+     *
+     */
+    private fun makePushId(uniqueId: Int) {
+        if (pushId == -1) {
+            pushId = uniqueId
         }
     }
 
-    private fun showNotification(context: Context, title: String, body: String) {
+    private fun showNotification(context: Context,
+                                 pushId: Int,
+                                 title: String,
+                                 body: String) {
         NotificationCreateHelper.showNotification(
                 context,
                 this,
+                pushId,
                 title,
                 body
         )
@@ -145,9 +177,26 @@ abstract class PushHandleStrategy<out T : BaseNotificationTypeData<*>> {
     abstract fun coldStartRoute(): ActivityRoute
 
     /**
-     * Метод для создания билдера ноификаций
+     * Метод для инициализации билдера нотификаций.
+     *
+     * @param context контекст
+     * @param title заголовок пуш-нотификации
+     * @param body текст пуш-нотификации
      */
-    abstract fun makeNotificationBuilder(context: Context, title: String, body: String): NotificationCompat.Builder?
+    abstract fun makeNotificationBuilder(context: Context,
+                                         title: String,
+                                         body: String): NotificationCompat.Builder?
+
+    /**
+     * Метод для инициализации builder'а заголовка группы нотификаций.
+     *
+     *
+     */
+    open fun makeGroupSummaryNotificationBuilder(context: Context,
+                                                 title: String,
+                                                 body: String): NotificationCompat.Builder? {
+        return null
+    }
 
     /**
      * Метод для создания канала нотификаций