+++ recycler-extension/src/main/java/ru/surfstudio/android/recycler/extension/sticky/layoutmanager/StickyItemPositioner.kt
@@ -700,7 +700,7 @@ class StickyItemPositioner(
     }
 
     companion object {
-        private const val INVALID_POSITION = -1
+        private val INVALID_POSITION = -1
     }
 
     enum class ElevationMode(val dp: Float) {