+++ shared-pref-sample/src/main/java/ru/surfstudio/android/shared/pref/sample/ui/screen/main/MainActivityView.kt
@@ -35,9 +35,11 @@ class MainActivityView : BaseLdsSwrActivityView<MainScreenModel>() {
 
     override fun getSwipeRefreshLayout(): SwipeRefreshLayout = swipe_refresh_layout
 
-    override fun onCreate(savedInstanceState: Bundle?,
-                          persistentState: PersistableBundle?,
-                          viewRecreated: Boolean) {
+    override fun onCreate(
+            savedInstanceState: Bundle?,
+            persistentState: PersistableBundle?,
+            viewRecreated: Boolean
+    ) {
         super.onCreate(savedInstanceState, persistentState, viewRecreated)
         initListeners()
     }
@@ -47,9 +49,9 @@ class MainActivityView : BaseLdsSwrActivityView<MainScreenModel>() {
         swipe_refresh_layout.goneIf(loadState != LoadState.NONE)
     }
 
-    override fun renderInternal(screenModel: MainScreenModel) {
-        ip_tv.text = screenModel.ip?.value
-        placeholder.render(screenModel.loadState)
+    override fun renderInternal(sm: MainScreenModel) {
+        ip_tv.text = sm.ip?.value
+        placeholder.render(sm.loadState)
     }
 
     private fun initListeners() {