+++ template/app-injector/src/main/java/ru/surfstudio/standard/app_injector/network/OkHttpModule.kt
@@ -1,5 +1,7 @@
 package ru.surfstudio.standard.app_injector.network
 
+import android.content.Context
+import com.readystatesoftware.chuck.ChuckInterceptor
 import ru.surfstudio.standard.i_network.service.ServiceInterceptor
 import ru.surfstudio.standard.i_token.TokenStorage
 import dagger.Module
@@ -10,6 +12,7 @@ import okhttp3.logging.HttpLoggingInterceptor
 import ru.surfstudio.android.dagger.scope.PerApplication
 import ru.surfstudio.android.network.cache.SimpleCacheInterceptor
 import ru.surfstudio.android.network.etag.EtagInterceptor
+import ru.surfstudio.standard.f_debug.server_settings.storage.DebugServerSettingsStorage
 import java.util.concurrent.TimeUnit
 import javax.inject.Named
 
@@ -35,13 +38,18 @@ class OkHttpModule {
             @Named(DI_NAME_SERVICE_INTERCEPTOR) serviceInterceptor: Interceptor,
             cacheInterceptor: SimpleCacheInterceptor,
             etagInterceptor: EtagInterceptor,
-            httpLoggingInterceptor: HttpLoggingInterceptor
+            httpLoggingInterceptor: HttpLoggingInterceptor,
+            debugServerSettingsStorage: DebugServerSettingsStorage,
+            context: Context
     ): OkHttpClient {
         return OkHttpClient.Builder().apply {
             connectTimeout(NETWORK_TIMEOUT, TimeUnit.SECONDS)
             readTimeout(NETWORK_TIMEOUT, TimeUnit.SECONDS)
             writeTimeout(NETWORK_TIMEOUT, TimeUnit.SECONDS)
 
+            if (debugServerSettingsStorage.isChuckEnabled) {
+                addInterceptor(ChuckInterceptor(context))
+            }
             addInterceptor(cacheInterceptor)
             addInterceptor(etagInterceptor)
             addInterceptor(serviceInterceptor)