+++ template/app-injector/src/main/java/ru/surfstudio/standard/app_injector/ui/ActivityComponent.kt
@@ -16,6 +16,10 @@ import ru.surfstudio.android.rx.extension.scheduler.SchedulersProvider
 import ru.surfstudio.android.rxbus.RxBus
 import ru.surfstudio.android.shared.pref.NO_BACKUP_SHARED_PREF
 import ru.surfstudio.standard.app_injector.AppComponent
+import ru.surfstudio.standard.f_debug.memory.storage.MemoryDebugStorage
+import ru.surfstudio.standard.f_debug.server_settings.reboot.interactor.RebootInteractor
+import ru.surfstudio.standard.f_debug.server_settings.storage.DebugServerSettingsStorage
+import ru.surfstudio.standard.i_initialization.InitializeAppInteractor
 import javax.inject.Named
 
 /**
@@ -29,6 +33,9 @@ interface ActivityComponent {
     fun connectionProvider(): ConnectionProvider
     fun activityProvider(): ActivityProvider
     fun stringsProvider(): StringsProvider
+    fun serverSettingsStorage(): DebugServerSettingsStorage
+    fun rebootInteractor(): RebootInteractor
+    fun memoryDebugStorage(): MemoryDebugStorage
 
     fun activityPersistentScope(): ActivityPersistentScope
     fun context(): Context
@@ -38,4 +45,6 @@ interface ActivityComponent {
     fun rxBus(): RxBus
     fun fcmStorage(): FcmStorage
     @Named(NO_BACKUP_SHARED_PREF) fun sharedPreferences(): SharedPreferences
+
+    fun initializeAppInteractor(): InitializeAppInteractor
 }