+++ template/f-debug/src/main/java/ru/surfstudio/standard/f_debug/common_controllers/CommonControllersDebugActivityView.kt
@@ -10,7 +10,7 @@ import ru.surfstudio.android.core.mvp.presenter.CorePresenter
 import ru.surfstudio.android.easyadapter.EasyAdapter
 import ru.surfstudio.android.easyadapter.ItemList
 import ru.surfstudio.android.template.f_debug.R
-import ru.surfstudio.standard.base_ui.component.provider.ComponentProvider
+import ru.surfstudio.standard.base_ui.provider.component.ComponentProvider
 import ru.surfstudio.standard.f_debug.common_controllers.controllers.CustomControllerDescriptionItemController
 import ru.surfstudio.standard.f_debug.common_controllers.description.addDescription
 import javax.inject.Inject
@@ -44,7 +44,7 @@ class CommonControllersDebugActivityView : BaseRenderableActivityView<CommonCont
         initAdapter()
     }
 
-    override fun renderInternal(screenModel: CommonControllersDebugScreenModel) {}
+    override fun renderInternal(sm: CommonControllersDebugScreenModel) {}
 
     override fun getScreenName(): String = "debug_controllers"