+++ template/f-main/src/main/java/ru/surfstudio/standard/f_main/MainActivityView.kt
@@ -3,11 +3,11 @@ package ru.surfstudio.standard.f_main
 import android.os.Bundle
 import android.os.PersistableBundle
 import androidx.annotation.LayoutRes
-import ru.surfstudio.standard.base_ui.component.provider.ComponentProvider
 import ru.surfstudio.android.core.mvp.activity.BaseRenderableActivityView
 import ru.surfstudio.android.core.mvp.presenter.CorePresenter
 import ru.surfstudio.android.core.ui.FragmentContainer
 import ru.surfstudio.android.template.f_main.R
+import ru.surfstudio.standard.base_ui.provider.component.ComponentProvider
 import javax.inject.Inject
 
 /**
@@ -38,7 +38,7 @@ class MainActivityView : BaseRenderableActivityView<MainScreenModel>(), Fragment
 
     override fun getScreenName(): String = "MainActivityView"
 
-    override fun renderInternal(screenModel: MainScreenModel) {
+    override fun renderInternal(sm: MainScreenModel) {
     }
 
     private fun initViews() {